pytorch/smp_baseline/requirements.txt
albumentations==1.1.0
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
entrypoints==0.4
executing==0.8.3
fonttools==4.33.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
imageio==2.17.0
imgaug==0.4.0
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.2.2
jupyter-core==4.10.0
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
munch==2.5.0
nest-asyncio==1.5.5
networkx==2.8
numpy==1.22.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pretrainedmodels==0.7.4
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.4
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@740dab561ccf54a9ae4bb5bda3b8b18df3790025
sentry-sdk==1.5.10
setproctitle==1.2.3
Shapely==1.8.1.post1
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
threadpoolctl==3.1.0
tifffile==2022.4.22
timm==0.4.12
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.15
wcwidth==0.2.5
